+++ src/controllers/directors.controllers.js
@@ -17,7 +17,11 @@ const get = async(req,res,next) => {
                 },
             ]
         });
-        return res.json(director);
+        if(director){
+            return res.json(director);
+        } else {
+            return res.status(404).json({message: `The director with id = ${id} does not exist`});
+        }
     }catch(error){
         next(error);
     }
@@ -25,8 +29,8 @@ const get = async(req,res,next) => {
 
 const getAll = async(req,res,next) => {
 
-    const limit = parseInt(req.query.limit);
-    const offset = parseInt(req.query.offset);
+    const limit = parseInt(req.query.limit) || 20;
+    const offset = parseInt(req.query.offset) || 0;
 
     try{
         let directors = await  Directors.findAll({
@@ -72,10 +76,13 @@ const deleteDirector = async(req,res,next) => {
     try{
         let director = await Directors.findOne({where: {id: id}});
 
-        await ContentDirectors.destroy({where: {director_id: id}});
-        await Directors.destroy({where: {id: id}});
-
-        return res.json(director);
+        if(director){
+            await ContentDirectors.destroy({where: {director_id: id}});
+            await Directors.destroy({where: {id: id}});
+            return res.json(director);
+        } else {
+            return res.status(204).json({message: `The director with id = ${id} does not exist`});
+        }
     }catch(error){
         next(error)
     }
@@ -107,7 +114,11 @@ const update = async(req,res,next) => {
                 },
             ]
         });
-        return res.json(director);
+        if(director){
+            return res.json(director);
+        } else {
+            return res.status(204).json({message: `The director with id = ${id} does not exist`});
+        }
     }catch(error){
         next(error)
     }
